Refactor iOS live smoke test support
